New Feature: User Locks
(Still needs some polish, but it seems to work pretty well.)

This feature adds a new tab, "Locks", where you can effectively disable or hide an upgrade to prevent it from being purchased by accident.
(It's mostly useful to e.g. suppress Shield upgrades in lategame, where the cost-per-unit of blocking power is astronomical)
The settings are preserved by a savegame, but cleared after a portal.